Refactor: files split.
